git-svn-id: https://darkengines.com/svn/darkengines/ima@454 302f0312-3c5d-0d4b-b714-136c201f073e

+++ ima/Precision.h
@@ -1,6 +1,6 @@
 #ifndef _PRECISION_H_
 #define _PRECISION_H_
 
-typedef double Real;
+typedef float Real;
 
 #endif
